backend/requirements.txt
asgiref==3.11.1
Django==5.2.12
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.17.0
gitdb==4.0.12
GitPython==3.1.41
smmap==5.0.2
sqlparse==0.5.5
